Fix AuthBin writing
Today I learned subtracting an int from a uint results in a long.

+++ src/TF3.YarhlPlugin.YakuzaKenzan/Converters/AuthBin/Writer.cs
@@ -100,7 +100,7 @@ namespace TF3.YarhlPlugin.YakuzaKenzan.Converters.AuthBin
             // Update file size
             reader.Stream.Seek(0x10);
             var fileSize = reader.ReadUInt32();
-            writer.Stream.RunInPosition(() => writer.Write(fileSize + sizeDifference), 0x10);
+            writer.Stream.RunInPosition(() => writer.Write((uint)(fileSize + sizeDifference)), 0x10);
 
             // Update offsets in file header
             var authHeaderOffsets = new uint[] { 0x14, 0x1C, 0x20, 0x30 };
@@ -109,7 +109,7 @@ namespace TF3.YarhlPlugin.YakuzaKenzan.Converters.AuthBin
                 reader.Stream.Seek(offset);
                 var offsetVal = reader.ReadUInt32();
                 if (offsetVal > nodeOffset)
-                    writer.Stream.RunInPosition(() => writer.Write(offsetVal + sizeDifference), offset);
+                    writer.Stream.RunInPosition(() => writer.Write((uint)(offsetVal + sizeDifference)), offset);
             }
 
             reader.Stream.Seek(0x1C);
@@ -130,7 +130,7 @@ namespace TF3.YarhlPlugin.YakuzaKenzan.Converters.AuthBin
                 reader.Stream.Seek(nodeStart + offset);
                 var offsetVal = reader.ReadUInt32();
                 if (offsetVal > nodeOffset)
-                    writer.Stream.RunInPosition(() => writer.Write(offsetVal + sizeDifference), nodeStart + offset);
+                    writer.Stream.RunInPosition(() => writer.Write((uint)(offsetVal + sizeDifference)), nodeStart + offset);
             }
 
             // Go to all subsequent nodes and update their offsets
